project and contractor creation
