+++ back/serializers/__init__.py
@@ -1,7 +1,7 @@
 from .auth_serializer import UserSerializer, UserSerializerWithToken  # noqa F401
 from .dashboard import DashboardSerializer  # noqa F401
-from .project import ProjectSerializer  # noqa F401
+from .project import ProjectSerializer, ProjectListSerializer  # noqa F401
 from .client_manager import ClientManagerSerializer  # noqa F401
-from .contractor import ContractorSerializer  # noqa F401
+from .contractor import ContractorSerializer, ContractorListSerializer  # noqa F401
 from .template import TemplateSerializer  # noqa F401
 from .worker import WorkerListSerializer, WorkerSerializer  # noqa F401
