+++ back/serializers/contractor.py
@@ -2,7 +2,13 @@ from rest_framework import serializers
 from back.models import Contractor
 
 
-class ContractorSerializer(serializers.ModelSerializer):
+class ContractorListSerializer(serializers.ModelSerializer):
     class Meta:
         model = Contractor
         fields = ["id", "username", "project_list"]
+
+
+class ContractorSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Contractor
+        fields = ["id", "username", "email"]
