+++ back/urls.py
@@ -27,10 +27,12 @@ urlpatterns = [
         r"password_reset/", reset_password_request_token, name="reset-password-request"
     ),
     path("projects/", views.ProjectList.as_view()),
+    path("projects/new/", views.ProjectCreate.as_view()),
     path("projects/<int:pk>/", views.ProjectDetail.as_view()),
     path("client_managers/", views.ClientManagerList.as_view()),
     path("client_managers/<int:pk>/", views.ClientManagerDetail.as_view()),
     path("contractors/", views.ContractorList.as_view()),
+    path("contractors/new/", views.ContractorCreate.as_view()),
     path("contractors/<int:pk>/", views.ContractorDetail.as_view()),
     path("templates/", views.TemplateList.as_view()),
     path("templates/<int:pk>/", views.TemplateDetail.as_view()),
