+++ back/views/__init__.py
@@ -1,7 +1,7 @@
 from .login import current_user, UserList, ObtainFoxJWTToken  # noqa F401
 from .dashboard import Dashboard  # noqa F401
-from .projects import ProjectList, ProjectDetail  # noqa F401
+from .projects import ProjectList, ProjectCreate, ProjectDetail  # noqa F401
 from .client_managers import ClientManagerList, ClientManagerDetail  # noqa F401
-from .contractors import ContractorList, ContractorDetail  # noqa F401
+from .contractors import ContractorList, ContractorDetail, ContractorCreate  # noqa F401
 from .templates import TemplateList, TemplateDetail  # noqa F401
 from .workers import WorkerList, WorkerDetail  # noqa F401
