+++ front/src/components/routes.js
@@ -6,6 +6,8 @@ const CurrentWorksList = React.lazy(() => import('./views/current_works/CurrentW
 const DocumentList = React.lazy(() => import('./views/documents/DocumentList'))
 const ClientManagerList = React.lazy(() => import('./views/managers/ClientManagerList'))
 const WorkerList = React.lazy(() => import('./views/workers/WorkerList'))
+const ProjectDetail = React.lazy(() => import('./views/projects/ProjectDetail'))
+const ContractorDetail = React.lazy(() => import('./views/contractors/ContractorDetail'))
 // const Toaster = React.lazy(() => import('./views/notifications/toaster/Toaster'));
 // const Tables = React.lazy(() => import('./views/base/tables/Tables'));
 
@@ -47,12 +49,15 @@ const WorkerList = React.lazy(() => import('./views/workers/WorkerList'))
 const routes = [
   { path: '/', exact: true, name: 'Fox' },
   { path: '/projects', name: 'Projects', component: ProjectList, exact: true },
+  { path: '/projects/new', name: 'New', component: ProjectDetail, exact: true },
   { path: '/approvals', name: 'Approvals', component: ApprovalList, exact: true },
   { path: '/current-works', name: 'Current Works', component: CurrentWorksList, exact: true },
   { path: '/contractors', name: 'Contractors', component: ContractorList, exact: true },
+  { path: '/contractors/new', name: 'New', component: ContractorDetail, exact: true },
   { path: '/documents', name: 'Documents', component: DocumentList, exact: true },
   { path: '/managers', name: 'Managers', component: ClientManagerList, exact: true },
   { path: '/workers', name: 'Workers', component: WorkerList, exact: true },
+
   // { path: '/dashboard', name: 'Dashboard', component: Dashboard },
   // { path: '/theme', name: 'Theme', component: Colors, exact: true },
   // { path: '/theme/colors', name: 'Colors', component: Colors },
@@ -95,3 +100,4 @@ const routes = [
 ];
 
 export default routes;
+
