+++ front/static/front/14.main.js
@@ -1,0 +1,15 @@
+(window["webpackJsonp"] = window["webpackJsonp"] || []).push([[14],{
+
+/***/ "./src/components/views/contractors/ContractorDetail.js":
+/*!**************************************************************!*\
+  !*** ./src/components/views/contractors/ContractorDetail.js ***!
+  \**************************************************************/
+/*! exports provided: default */
+/***/ (function(module, exports) {
+
+throw new Error("Module build failed (from ./node_modules/babel-loader/lib/index.js):\nSyntaxError: D:\\simple2b\\FOX\\front\\src\\components\\views\\contractors\\ContractorDetail.js: Unexpected token (104:8)\n\n\u001b[0m \u001b[90m 102 | \u001b[39m  }\u001b[0m\n\u001b[0m \u001b[90m 103 | \u001b[39m\u001b[0m\n\u001b[0m\u001b[31m\u001b[1m>\u001b[22m\u001b[39m\u001b[90m 104 | \u001b[39m  \u001b[36mconst\u001b[39m mapStateToProps \u001b[33m=\u001b[39m state \u001b[33m=>\u001b[39m {\u001b[0m\n\u001b[0m \u001b[90m     | \u001b[39m        \u001b[31m\u001b[1m^\u001b[22m\u001b[39m\u001b[0m\n\u001b[0m \u001b[90m 105 | \u001b[39m    \u001b[36mreturn\u001b[39m {\u001b[0m\n\u001b[0m \u001b[90m 106 | \u001b[39m      company\u001b[33m:\u001b[39m state\u001b[33m.\u001b[39mcurrentUser\u001b[33m.\u001b[39mcompany\u001b[33m,\u001b[39m\u001b[0m\n\u001b[0m \u001b[90m 107 | \u001b[39m      options\u001b[33m:\u001b[39m state\u001b[33m.\u001b[39mentityListTable\u001b[33m.\u001b[39mtableData\u001b[0m\n    at Object._raise (D:\\simple2b\\FOX\\front\\node_modules\\@babel\\parser\\lib\\index.js:757:17)\n    at Object.raiseWithData (D:\\simple2b\\FOX\\front\\node_modules\\@babel\\parser\\lib\\index.js:750:17)\n    at Object.raise (D:\\simple2b\\FOX\\front\\node_modules\\@babel\\parser\\lib\\index.js:744:17)\n    at Object.unexpected (D:\\simple2b\\FOX\\front\\node_modules\\@babel\\parser\\lib\\index.js:8834:16)\n    at Object.parseClassMemberWithIsStatic (D:\\simple2b\\FOX\\front\\node_modules\\@babel\\parser\\lib\\index.js:12176:12)\n    at Object.parseClassMember (D:\\simple2b\\FOX\\front\\node_modules\\@babel\\parser\\lib\\index.js:12069:10)\n    at D:\\simple2b\\FOX\\front\\node_modules\\@babel\\parser\\lib\\index.js:12014:14\n    at Object.withTopicForbiddingContext (D:\\simple2b\\FOX\\front\\node_modules\\@babel\\parser\\lib\\index.js:11085:14)\n    at Object.parseClassBody (D:\\simple2b\\FOX\\front\\node_modules\\@babel\\parser\\lib\\index.js:11991:10)\n    at Object.parseClass (D:\\simple2b\\FOX\\front\\node_modules\\@babel\\parser\\lib\\index.js:11965:22)\n    at Object.parseStatementContent (D:\\simple2b\\FOX\\front\\node_modules\\@babel\\parser\\lib\\index.js:11252:21)\n    at Object.parseStatement (D:\\simple2b\\FOX\\front\\node_modules\\@babel\\parser\\lib\\index.js:11210:17)\n    at Object.parseBlockOrModuleBlockBody (D:\\simple2b\\FOX\\front\\node_modules\\@babel\\parser\\lib\\index.js:11785:25)\n    at Object.parseBlockBody (D:\\simple2b\\FOX\\front\\node_modules\\@babel\\parser\\lib\\index.js:11771:10)\n    at Object.parseTopLevel (D:\\simple2b\\FOX\\front\\node_modules\\@babel\\parser\\lib\\index.js:11141:10)\n    at Object.parse (D:\\simple2b\\FOX\\front\\node_modules\\@babel\\parser\\lib\\index.js:12843:10)\n    at parse (D:\\simple2b\\FOX\\front\\node_modules\\@babel\\parser\\lib\\index.js:12896:38)\n    at parser (D:\\simple2b\\FOX\\front\\node_modules\\@babel\\core\\lib\\parser\\index.js:54:34)\n    at parser.next (<anonymous>)\n    at normalizeFile (D:\\simple2b\\FOX\\front\\node_modules\\@babel\\core\\lib\\transformation\\normalize-file.js:93:38)\n    at normalizeFile.next (<anonymous>)\n    at run (D:\\simple2b\\FOX\\front\\node_modules\\@babel\\core\\lib\\transformation\\index.js:31:50)\n    at run.next (<anonymous>)\n    at Function.transform (D:\\simple2b\\FOX\\front\\node_modules\\@babel\\core\\lib\\transform.js:27:41)\n    at transform.next (<anonymous>)\n    at step (D:\\simple2b\\FOX\\front\\node_modules\\gensync\\index.js:254:32)\n    at D:\\simple2b\\FOX\\front\\node_modules\\gensync\\index.js:266:13\n    at async.call.result.err.err (D:\\simple2b\\FOX\\front\\node_modules\\gensync\\index.js:216:11)");
+
+/***/ })
+
+}]);
+//# sourceMappingURL=14.main.js.map
