+++ front/static/front/3.main.js
@@ -1,9 +1,9 @@
 (window["webpackJsonp"] = window["webpackJsonp"] || []).push([[3],{
 
-/***/ "./src/components/cards/FoxRegisterCard.js":
-/*!*************************************************!*\
-  !*** ./src/components/cards/FoxRegisterCard.js ***!
-  \*************************************************/
+/***/ "./src/components/cards/FoxLoginCard.js":
+/*!**********************************************!*\
+  !*** ./src/components/cards/FoxLoginCard.js ***!
+  \**********************************************/
 /*! exports provided: default */
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
@@ -12,7 +12,7 @@ __webpack_require__.r(__webpack_exports__);
 /* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react */ "./node_modules/react/index.js");
 /* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(react__WEBPACK_IMPORTED_MODULE_0__);
 /* harmony import */ var _coreui_react__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @coreui/react */ "./node_modules/@coreui/react/es/index.js");
-/* harmony import */ var _forms_FoxRegisterForm__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../forms/FoxRegisterForm */ "./src/components/forms/FoxRegisterForm.js");
+/* harmony import */ var _forms_FoxLoginForm__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../forms/FoxLoginForm */ "./src/components/forms/FoxLoginForm.js");
 function _typeof(obj) { "@babel/helpers - typeof"; if (typeof Symbol === "function" && typeof Symbol.iterator === "symbol") { _typeof = function _typeof(obj) { return typeof obj; }; } else { _typeof = function _typeof(obj) { return obj && typeof Symbol === "function" && obj.constructor === Symbol && obj !== Symbol.prototype ? "symbol" : typeof obj; }; } return _typeof(obj); }
 
 function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
@@ -39,40 +39,38 @@ function _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.g
 
 
 
-var FoxRegisterCard = /*#__PURE__*/function (_Component) {
-  _inherits(FoxRegisterCard, _Component);
+var FoxLoginCard = /*#__PURE__*/function (_Component) {
+  _inherits(FoxLoginCard, _Component);
 
-  var _super = _createSuper(FoxRegisterCard);
+  var _super = _createSuper(FoxLoginCard);
 
-  function FoxRegisterCard() {
-    _classCallCheck(this, FoxRegisterCard);
+  function FoxLoginCard() {
+    _classCallCheck(this, FoxLoginCard);
 
     return _super.apply(this, arguments);
   }
 
-  _createClass(FoxRegisterCard, [{
+  _createClass(FoxLoginCard, [{
     key: "render",
     value: function render() {
       return /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(_coreui_react__WEBPACK_IMPORTED_MODULE_1__["CCard"], null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(_coreui_react__WEBPACK_IMPORTED_MODULE_1__["CCardHeader"], {
         color: "dark",
         textColor: "light"
-      }, "Register"), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(_forms_FoxRegisterForm__WEBPACK_IMPORTED_MODULE_2__["default"], {
-        username: this.props.username
-      }));
+      }, "Login"), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(_forms_FoxLoginForm__WEBPACK_IMPORTED_MODULE_2__["default"], null));
     }
   }]);
 
-  return FoxRegisterCard;
+  return FoxLoginCard;
 }(react__WEBPACK_IMPORTED_MODULE_0__["Component"]);
 
-/* harmony default export */ __webpack_exports__["default"] = (FoxRegisterCard);
+/* harmony default export */ __webpack_exports__["default"] = (FoxLoginCard);
 
 /***/ }),
 
-/***/ "./src/components/forms/FoxRegisterForm.js":
-/*!*************************************************!*\
-  !*** ./src/components/forms/FoxRegisterForm.js ***!
-  \*************************************************/
+/***/ "./src/components/forms/FoxLoginForm.js":
+/*!**********************************************!*\
+  !*** ./src/components/forms/FoxLoginForm.js ***!
+  \**********************************************/
 /*! exports provided: default */
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
@@ -115,15 +113,15 @@ function _defineProperty(obj, key, value) { if (key in obj) { Object.definePrope
 
 
 
-var FoxRegisterForm = /*#__PURE__*/function (_Component) {
-  _inherits(FoxRegisterForm, _Component);
+var FoxLoginForm = /*#__PURE__*/function (_Component) {
+  _inherits(FoxLoginForm, _Component);
 
-  var _super = _createSuper(FoxRegisterForm);
+  var _super = _createSuper(FoxLoginForm);
 
-  function FoxRegisterForm() {
+  function FoxLoginForm() {
     var _this;
 
-    _classCallCheck(this, FoxRegisterForm);
+    _classCallCheck(this, FoxLoginForm);
 
     for (var _len = arguments.length, args = new Array(_len), _key = 0; _key < _len; _key++) {
       args[_key] = arguments[_key];
@@ -132,39 +130,26 @@ var FoxRegisterForm = /*#__PURE__*/function (_Component) {
     _this = _super.call.apply(_super, [this].concat(args));
 
     _defineProperty(_assertThisInitialized(_this), "state", {
-      username: _this.props.username,
-      password: "",
-      password2: ""
+      username: "",
+      password: ""
     });
 
     _defineProperty(_assertThisInitialized(_this), "handleChange", function (event) {
-      _this.setState(_defineProperty({}, event.target.name, event.target.value), function () {
-        if (_this.state.password !== _this.state.password2) {
-          _this.props.registerError();
-        } else {
-          _this.props.allowRegistration(_this.props.registrationToken);
-        }
-
-        ;
-      });
+      _this.setState(_defineProperty({}, event.target.name, event.target.value));
     });
 
     _defineProperty(_assertThisInitialized(_this), "handleSubmit", function (event) {
       event.preventDefault();
 
-      _this.props.userRegisterFetch({
-        password: _this.state.password,
-        token: _this.props.registrationToken
-      });
+      _this.props.userLoginFetch(_this.state);
     });
 
     return _this;
   }
 
-  _createClass(FoxRegisterForm, [{
+  _createClass(FoxLoginForm, [{
     key: "render",
     value: function render() {
-      console.log(this.props.errorMessage);
       return /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(_coreui_react__WEBPACK_IMPORTED_MODULE_2__["CContainer"], {
         fluid: true
       }, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(_coreui_react__WEBPACK_IMPORTED_MODULE_2__["CRow"], {
@@ -182,7 +167,6 @@ var FoxRegisterForm = /*#__PURE__*/function (_Component) {
         placeholder: "Enter your name",
         value: this.state.username,
         onChange: this.handleChange,
-        readOnly: true,
         required: true
       })), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(_coreui_react__WEBPACK_IMPORTED_MODULE_2__["CFormGroup"], null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(_coreui_react__WEBPACK_IMPORTED_MODULE_2__["CLabel"], {
         htmlFor: "nf-password"
@@ -194,58 +178,40 @@ var FoxRegisterForm = /*#__PURE__*/function (_Component) {
         value: this.state.password,
         onChange: this.handleChange,
         required: true
-      }), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(_coreui_react__WEBPACK_IMPORTED_MODULE_2__["CLabel"], {
-        htmlFor: "nf-password"
-      }, "Confirm Password"), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(_coreui_react__WEBPACK_IMPORTED_MODULE_2__["CInput"], {
-        type: "password",
-        id: "password2",
-        name: "password2",
-        placeholder: "Confirm Password..",
-        value: this.state.password2,
-        onChange: this.handleChange,
-        required: true
       })), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(_coreui_react__WEBPACK_IMPORTED_MODULE_2__["CFormGroup"], null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(_coreui_react__WEBPACK_IMPORTED_MODULE_2__["CInput"], {
         type: "submit",
         value: "Submit",
-        color: "info",
-        disabled: this.state.password !== this.state.password2
-      })), this.props.registerErrorFlag ? /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement("p", null, this.props.errorMessage ? this.props.errorMessage : 'INVALID CREDENTIALS! PLEASE, CHECK YOUR PASSWORD AND PASSWORD CONFIRMATION FIELDS!') : null))));
+        color: "info"
+      })), this.props.loginError ? /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement("p", null, "INVALID CREDENTIALS! PLEASE, CHECK YOUR USERNAME AND PASSWORD!") : null))));
     }
   }]);
 
-  return FoxRegisterForm;
+  return FoxLoginForm;
 }(react__WEBPACK_IMPORTED_MODULE_0__["Component"]);
 
 var mapStateToProps = function mapStateToProps(state) {
   return {
-    registerErrorFlag: state.registerError,
-    registrationToken: state.registrationToken,
-    errorMessage: state.errorMessage
+    currentUser: state.currentUser,
+    loginError: state.loginError
   };
 };
 
 var mapDispatchToProps = function mapDispatchToProps(dispatch) {
   return {
-    userRegisterFetch: function userRegisterFetch(credentials) {
-      return dispatch(Object(_actions__WEBPACK_IMPORTED_MODULE_4__["userRegisterFetch"])(credentials));
-    },
-    registerError: function registerError() {
-      return dispatch(Object(_actions__WEBPACK_IMPORTED_MODULE_4__["registerError"])());
-    },
-    allowRegistration: function allowRegistration(token) {
-      return dispatch(Object(_actions__WEBPACK_IMPORTED_MODULE_4__["allowRegistration"])(token));
+    userLoginFetch: function userLoginFetch(userInfo) {
+      return dispatch(Object(_actions__WEBPACK_IMPORTED_MODULE_4__["userLoginFetch"])(userInfo));
     }
   };
 };
 
-/* harmony default export */ __webpack_exports__["default"] = (Object(react_redux__WEBPACK_IMPORTED_MODULE_1__["connect"])(mapStateToProps, mapDispatchToProps)(FoxRegisterForm));
+/* harmony default export */ __webpack_exports__["default"] = (Object(react_redux__WEBPACK_IMPORTED_MODULE_1__["connect"])(mapStateToProps, mapDispatchToProps)(FoxLoginForm));
 
 /***/ }),
 
-/***/ "./src/components/pages/Register.js":
-/*!******************************************!*\
-  !*** ./src/components/pages/Register.js ***!
-  \******************************************/
+/***/ "./src/components/pages/Login.js":
+/*!***************************************!*\
+  !*** ./src/components/pages/Login.js ***!
+  \***************************************/
 /*! exports provided: default */
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
@@ -254,112 +220,22 @@ __webpack_require__.r(__webpack_exports__);
 /* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react */ "./node_modules/react/index.js");
 /* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(react__WEBPACK_IMPORTED_MODULE_0__);
 /* harmony import */ var _coreui_react__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @coreui/react */ "./node_modules/@coreui/react/es/index.js");
-/* harmony import */ var react_redux__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! react-redux */ "./node_modules/react-redux/es/index.js");
-/* harmony import */ var react_router_dom__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! react-router-dom */ "./node_modules/react-router-dom/esm/react-router-dom.js");
-/* harmony import */ var _cards_FoxRegisterCard__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ../cards/FoxRegisterCard */ "./src/components/cards/FoxRegisterCard.js");
-/* harmony import */ var _actions__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! ../../actions */ "./src/actions/index.js");
-function _typeof(obj) { "@babel/helpers - typeof"; if (typeof Symbol === "function" && typeof Symbol.iterator === "symbol") { _typeof = function _typeof(obj) { return typeof obj; }; } else { _typeof = function _typeof(obj) { return obj && typeof Symbol === "function" && obj.constructor === Symbol && obj !== Symbol.prototype ? "symbol" : typeof obj; }; } return _typeof(obj); }
-
-function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
-
-function _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }
-
-function _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); return Constructor; }
-
-function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); if (superClass) _setPrototypeOf(subClass, superClass); }
-
-function _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf || function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }
-
-function _createSuper(Derived) { var hasNativeReflectConstruct = _isNativeReflectConstruct(); return function _createSuperInternal() { var Super = _getPrototypeOf(Derived), result; if (hasNativeReflectConstruct) { var NewTarget = _getPrototypeOf(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return _possibleConstructorReturn(this, result); }; }
-
-function _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === "object" || typeof call === "function")) { return call; } return _assertThisInitialized(self); }
-
-function _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return self; }
-
-function _isNativeReflectConstruct() { if (typeof Reflect === "undefined" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === "function") return true; try { Date.prototype.toString.call(Reflect.construct(Date, [], function () {})); return true; } catch (e) { return false; } }
-
-function _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }
-
-function _defineProperty(obj, key, value) { if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }
+/* harmony import */ var _cards_FoxLoginCard__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../cards/FoxLoginCard */ "./src/components/cards/FoxLoginCard.js");
 
 
 
 
-
-
-
-var loading = /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement("div", {
-  className: "pt-3 text-center"
-}, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(_coreui_react__WEBPACK_IMPORTED_MODULE_1__["CSpinner"], {
-  size: "sm",
-  variant: "grow",
-  style: {
-    width: '4rem',
-    height: '4rem'
-  }
-}));
-
-var Register = /*#__PURE__*/function (_Component) {
-  _inherits(Register, _Component);
-
-  var _super = _createSuper(Register);
-
-  function Register() {
-    var _this;
-
-    _classCallCheck(this, Register);
-
-    for (var _len = arguments.length, args = new Array(_len), _key = 0; _key < _len; _key++) {
-      args[_key] = arguments[_key];
-    }
-
-    _this = _super.call.apply(_super, [this].concat(args));
-
-    _defineProperty(_assertThisInitialized(_this), "componentDidMount", function () {
-      _this.props.userRegisterValidationFetch(_this.props.token);
-    });
-
-    return _this;
-  }
-
-  _createClass(Register, [{
-    key: "render",
-    value: function render() {
-      return !this.props.registerAllowed ? /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(react_router_dom__WEBPACK_IMPORTED_MODULE_3__["Redirect"], {
-        to: "/login"
-      }) : /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(react__WEBPACK_IMPORTED_MODULE_0__["Suspense"], {
-        fallback: loading
-      }, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement("div", {
-        className: "c-app c-default-layout flex-row align-items-center"
-      }, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(_coreui_react__WEBPACK_IMPORTED_MODULE_1__["CContainer"], null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(_coreui_react__WEBPACK_IMPORTED_MODULE_1__["CRow"], {
-        className: "justify-content-center"
-      }, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(_coreui_react__WEBPACK_IMPORTED_MODULE_1__["CCol"], {
-        md: "6"
-      }, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(_cards_FoxRegisterCard__WEBPACK_IMPORTED_MODULE_4__["default"], {
-        username: this.props.username
-      }))))));
-    }
-  }]);
-
-  return Register;
-}(react__WEBPACK_IMPORTED_MODULE_0__["Component"]);
-
-var mapStateToProps = function mapStateToProps(state) {
-  return {
-    userRegistered: state.userRegistered,
-    registerAllowed: state.registerAllowed
-  };
+var Login = function Login() {
+  return /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement("div", {
+    className: "c-app c-default-layout flex-row align-items-center"
+  }, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(_coreui_react__WEBPACK_IMPORTED_MODULE_1__["CContainer"], null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(_coreui_react__WEBPACK_IMPORTED_MODULE_1__["CRow"], {
+    className: "justify-content-center"
+  }, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(_coreui_react__WEBPACK_IMPORTED_MODULE_1__["CCol"], {
+    md: "6"
+  }, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(_cards_FoxLoginCard__WEBPACK_IMPORTED_MODULE_2__["default"], null)))));
 };
 
-var mapDispatchToProps = function mapDispatchToProps(dispatch) {
-  return {
-    userRegisterValidationFetch: function userRegisterValidationFetch(token) {
-      return dispatch(Object(_actions__WEBPACK_IMPORTED_MODULE_5__["userRegisterValidationFetch"])(token));
-    }
-  };
-};
-
-/* harmony default export */ __webpack_exports__["default"] = (Object(react_redux__WEBPACK_IMPORTED_MODULE_2__["connect"])(mapStateToProps, mapDispatchToProps)(Register));
+/* harmony default export */ __webpack_exports__["default"] = (Login);
 
 /***/ })
